agregando validaciones

+++ src/routes/usersRouter.js
@@ -14,13 +14,19 @@ router.get("/", async (req, res, next) => {
 });
 
 router.post("/register", async (req, res, next) => {
-    const { first_name, last_name, email, role, password } = req.body;
-
+    const { first_name, last_name, email, age, role, password } = req.body;
+    if (!first_name || !last_name || !email || !age || !role || !password) {
+        return res.status(400).json({ message: "All fields are required" });
+    }
+    if (await UserModel.findOne({ email })) {
+        return res.status(400).json({ message: "User already exists" });
+    }
     try {
         const user = await UserModel.create({
             first_name,
             last_name,
             email,
+            age,
             role,
             password: createHash(password),
         });
@@ -32,8 +38,18 @@ router.post("/register", async (req, res, next) => {
 
 router.put("/update/:id", async (req, res, next) => {
     const { first_name, last_name, password } = req.body;
+    if (!first_name || !last_name || !password) {
+        return res.status(400).json({ message: "All fields are required" });
+    }
+    if (await UserModel.findOne({ email })) {
+        return res.status(400).json({ message: "User already exists" });
+    }
     try {
-        const user = await UserModel.findByIdAndUpdate(req.params.id,req.body,{ new: true });
+        const user = await UserModel.findByIdAndUpdate(
+            req.params.id,
+            req.body,
+            { new: true }
+        );
         res.status(200).json(user);
     } catch (error) {
         res.status(500).json({ message: error.message });
@@ -42,6 +58,9 @@ router.put("/update/:id", async (req, res, next) => {
 
 router.delete("/delete/:id", async (req, res, next) => {
     const { id } = req.params;
+    if (!id) {
+        return res.status(400).json({ message: "Id is required" });
+    }
     try {
         await UserModel.findByIdAndDelete(req.params.id);
         res.status(200).json({ message: "User deleted successfully" });
